+++ RexIntents/RexIntents.swift
@@ -71,37 +71,3 @@ extension MatchMode: AppEnum {
         .all: "all matches",
     ]
 }
-
-public struct RegexEntity {
-    
-    
-    public let id: UUID
-    public let name: String
-}
-
-// MARK: - AppEntity Conformance
-extension RegexEntity: AppEntity {
-    
-    public static var typeDisplayRepresentation: TypeDisplayRepresentation {
-        "Regular Expression"
-    }
-    
-    public var displayRepresentation: DisplayRepresentation {
-        DisplayRepresentation(title: "\(name)")
-    }
-    
-    public static var defaultQuery: RegexQuery {
-        .init()
-    }
-}
-
-public struct RegexQuery: EntityQuery {
-    
-    public typealias Entity = RegexEntity
-    
-    public init() { }
-    
-    public func entities(for identifiers: [Entity.ID]) async throws -> [Entity] {
-        return []
-    }
-}
